20_report_craft_partially/requirements.txt
# 귀책분석 자동화 시스템 — 의존성
# 설치: pip install -r requirements.txt

# ── 문서 생성 ────────────────────────────────────────────────────
python-docx>=1.1.0          # Word 문서 읽기/생성

# ── PDF 텍스트 추출 ──────────────────────────────────────────────
pdfplumber>=0.10.0          # PDF 텍스트 추출 (기본)
PyMuPDF>=1.23.0             # PDF 텍스트/이미지 추출 (폴백)

# ── OCR ─────────────────────────────────────────────────────────
pytesseract>=0.3.10         # OCR (Tesseract 엔진 별도 설치 필요)
Pillow>=10.0.0              # 이미지 처리

# ── 진행바 ──────────────────────────────────────────────────────
tqdm>=4.66.0                # 진행 상태 표시

# ── 환경변수 ─────────────────────────────────────────────────────
python-dotenv>=1.0.0        # .env 파일 로드

# ── 선택: HWP 지원 ───────────────────────────────────────────────
# hwp5==0.1.0               # python-hwp (설치 실패 시 LibreOffice 폴백 사용)
